requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.8
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.8
